# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the Ridash API server for user authentication and management",
        "title": "Ridash API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/teams/{teamID}/status-pages/{id}": {
            "get": {
                "description": "Fetches a status page with groups and monitors for a team the user belongs to",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "status_pages"
                ],
                "summary": "Get a status page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status Page ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status page retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid team ID or status page ID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Status page not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates slug/icon/groups/monitors for the given status page (owner/admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "status_pages"
                ],
                "summary": "Update a status page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status Page ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status page update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/statuspage.statusPageUpsertRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a status page for a team (owner/admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "status-pages"
                ],
                "summary": "Delete a status page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status Page ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status page deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid team ID or status page ID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Status page not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.StatusPageElementType": {
            "type": "string",
            "enum": [
                "historical_timeline",
                "current_status_indicator"
            ],
            "x-enum-varnames": [
                "StatusPageElementTypeHistoricalTimeline",
                "StatusPageElementTypeCurrentStatusIndicator"
            ]
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Invalid request body"
                }
            }
        },
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                }
            }
        },
        "statuspage.statusPageElementInput": {
            "type": "object",
            "required": [
                "name",
                "type"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "monitor": {
                    "type": "boolean"
                },
                "monitor_id": {
                    "type": "string",
                    "example": "0"
                },
                "monitors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/statuspage.statusPageMonitorInput"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "sort_order": {
                    "type": "integer",
                    "minimum": 1
                },
                "type": {
                    "enum": [
                        "historical_timeline",
                        "current_status_indicator"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.StatusPageElementType"
                        }
                    ]
                }
            }
        },
        "statuspage.statusPageGroupInput": {
            "type": "object",
            "required": [
                "name",
                "type"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "sort_order": {
                    "type": "integer",
                    "minimum": 1
                },
                "type": {
                    "enum": [
                        "historical_timeline",
                        "current_status_indicator"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.StatusPageElementType"
                        }
                    ]
                }
            }
        },
        "statuspage.statusPageMonitorInput": {
            "type": "object",
            "required": [
                "monitor_id",
                "name",
                "type"
            ],
            "properties": {
                "group_id": {
                    "type": "string",
                    "example": "0"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "monitor_id": {
                    "type": "string",
                    "example": "0"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "sort_order": {
                    "type": "integer",
                    "minimum": 1
                },
                "type": {
                    "enum": [
                        "historical_timeline",
                        "current_status_indicator"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.StatusPageElementType"
                        }
                    ]
                }
            }
        },
        "statuspage.statusPageUpsertRequest": {
            "type": "object",
            "required": [
                "slug",
                "title"
            ],
            "properties": {
                "elements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/statuspage.statusPageElementInput"
                    }
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/statuspage.statusPageGroupInput"
                    }
                },
                "icon": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "monitors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/statuspage.statusPageMonitorInput"
                    }
                },
                "slug": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}